ai-model/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
numpy==1.26.4
tensorflow==2.16.2
scikit-learn>=1.3.0
